update unit test namespaces
