+++ tests/OrasProject.Oras.Tests/Registry/Remote/Auth/CacheTest.cs
