+++ tests/OrasProject.Oras.Tests/Registry/Remote/Auth/ChallengeTest.cs
