+++ tests/OrasProject.Oras.Tests/Registry/Remote/Auth/ClientTest.cs
