+++ tests/OrasProject.Oras.Tests/Registry/Remote/Auth/ScopeManagerTest.cs
