+++ tests/OrasProject.Oras.Tests/Registry/Remote/Auth/ScopeTest.cs
