+++ tests/OrasProject.Oras.Tests/Registry/Remote/Auth/SingleRegistryCredentialProviderTests.cs
