+++ tests/OrasProject.Oras.Tests/Registry/Remote/HttpRequestMessageExtensionsTest.cs
