+++ tests/OrasProject.Oras.Tests/Registry/Remote/HttpResponseExceptionTests.cs
