+++ tests/OrasProject.Oras.Tests/Registry/Remote/NonSeekableStream.cs
