+++ tests/OrasProject.Oras.Tests/Registry/Remote/ReferenceTest.cs
