+++ tests/OrasProject.Oras.Tests/Registry/Remote/UriFactoryTest.cs
