+++ tests/OrasProject.Oras.Tests/Registry/Remote/Util/RandomDataGenerator.cs
